Replace AdressServiceTest with AddressServiceIntegrationTest
